fixup! crypto-sodium: Add blake2b hash
